Split admin insights page helpers
